+++ src/srcml/srcml_nodes.hpp
@@ -9,7 +9,7 @@
 typedef std::vector<std::shared_ptr<srcml_node>> srcml_nodes;
 
 void advance_to_child(const srcml_nodes & nodes,
-                      unsigned int & pos,
+                      size_t & pos,
                       xmlElementType type,
                       const std::string & name);
